add jpa repositories

+++ src/main/java/com/ninjarmm/repositories/DeviceTypeRepository.java
@@ -1,9 +1,10 @@
 package com.ninjarmm.repositories;
 
 import com.ninjarmm.entities.DeviceType;
+import io.swagger.models.auth.In;
 import org.springframework.data.repository.CrudRepository;
 
-public interface DeviceTypeRepository extends CrudRepository<DeviceType, Long>{
+public interface DeviceTypeRepository extends CrudRepository<DeviceType, Integer>{
 
-  DeviceType findById(long id);
+  DeviceType findById(Integer id);
 }
